Attempt to move threading-related code into submodule
