+++ crates/kernel/examples/main.rs
@@ -5,6 +5,7 @@ extern crate alloc;
 extern crate kernel;
 
 use kernel::*;
+use kernel::event::thread;
 
 #[no_mangle]
 extern "Rust" fn kernel_main(_device_tree: device_tree::DeviceTree) {
