+++ crates/kernel/src/event/task.rs
@@ -5,14 +5,14 @@ use core::pin::Pin;
 use core::sync::atomic::{AtomicUsize, Ordering};
 use core::task::{Context, Poll, RawWaker, RawWakerVTable, Waker};
 
-use crate::event::SCHEDULER;
+use super::SCHEDULER;
 use crate::sync::SpinLock;
 
 pub fn spawn_async(future: impl Future<Output = ()> + Send + Sync + 'static) {
     let task = TASKS.add_task(Task {
         future: Box::pin(future),
     });
-    SCHEDULER.add_task(crate::event::Event::AsyncTask(task));
+    SCHEDULER.add_task(super::Event::AsyncTask(task));
 }
 
 pub static TASKS: TaskList = TaskList::new();
@@ -101,7 +101,7 @@ impl TaskList {
 pub struct TaskId(usize);
 
 fn wake_task(task: TaskId) {
-    SCHEDULER.add_task(crate::event::Event::AsyncTask(task));
+    SCHEDULER.add_task(super::Event::AsyncTask(task));
 }
 
 static WAKER_VTABLE: RawWakerVTable = RawWakerVTable::new(
