+++ crates/kernel/src/syscall/sync.rs
@@ -1,4 +1,4 @@
-use crate::context::{deschedule_thread, Context, DescheduleAction, CORES};
+use crate::event::context::{deschedule_thread, Context, DescheduleAction, CORES};
 
 pub unsafe fn sys_yield(ctx: &mut Context) -> *mut Context {
     let (core_sp, thread) = CORES.with_current(|core| (core.core_sp.get(), core.thread.take()));
